backend/requirements.txt
fastapi
uvicorn[standard]
httpx
ddgs
python-dotenv